Added security to views

+++ products/views.py
@@ -1,5 +1,6 @@
 from django.shortcuts import render, redirect, reverse, get_object_or_404
 from django.contrib import messages
+from django.contrib.auth.decorators import login_required
 from django.db.models import Q
 from django.db.models.functions import Lower
 from .models import Product, Category
@@ -72,8 +73,12 @@ def product_detail(request, slug):
     return render(request, 'products/product_detail.html', context)
 
 
+@login_required
 def add_product(request):
     """ Add a product to the store """
+    if not request.user.is_superuser:
+        messages.error(request, 'Sorry! Only store owners are allowed here.')
+        return redirect(reverse('home'))
     if request.method == 'POST':
         form = ProductForm(request.POST, request.FILES)
         if form.is_valid():
@@ -94,8 +99,13 @@ def add_product(request):
     return render(request, template, context)
 
 
+@login_required
 def edit_product(request, slug):
     """ Edit a product in the store """
+    if not request.user.is_superuser:
+        messages.error(request, 'Sorry! Only store owners are allowed here.')
+        return redirect(reverse('home'))
+        
     product = get_object_or_404(Product, slug=slug)
     if request.method == 'POST':
         form = ProductForm(request.POST, request.FILES, instance=product)
@@ -118,8 +128,13 @@ def edit_product(request, slug):
     return render(request, template, context)
 
 
+@login_required
 def delete_product(request, slug):
     """ Delete a product """
+    if not request.user.is_superuser:
+        messages.error(request, 'Sorry! Only store owners are allowed here.')
+        return redirect(reverse('home'))
+
     product = get_object_or_404(Product, slug=slug)
     product.delete()
     messages.success(request, f'{product.name} deleted!')
